Node

Description

Defines the set of dataflow node constructors used by the builder/codegen stages and utilities to query node names, arity, and standard output ports.

Documentation

data DNode

Dataflow node (mirrors assembler mnemonics).

Constructors

NConstI

Fields

nName :: !String

cInt :: !Int

NConstF

Fields

nName :: !String

cFloat :: !Float

NConstD

Fields

nName :: !String

cDouble :: !Double

NAdd

Fields

nName :: !String

NSub

Fields

nName :: !String

NMul

Fields

nName :: !String

NDiv
2 outputs

Fields

nName :: !String

NFAdd

Fields

nName :: !String

NDAdd

Fields

nName :: !String

NBand

Fields

nName :: !String

NFSub
new

Fields

nName :: !String

NFMul
new

Fields

nName :: !String

NFDiv
new

Fields

nName :: !String

NAddI

Fields

nName :: !String

iImm :: !Int

NSubI

Fields

nName :: !String

iImm :: !Int

NMulI

Fields

nName :: !String

iImm :: !Int

NFMulI

Fields

nName :: !String

fImm :: !Float

NDivI
2 outputs

Fields

nName :: !String

iImm :: !Int

NLThan

Fields

nName :: !String

NGThan

Fields

nName :: !String

NEqual

Fields

nName :: !String

NLThanI

Fields

nName :: !String

iImm :: !Int

NGThanI

Fields

nName :: !String

iImm :: !Int

NSteer
ports "t" / "f"

Fields

nName :: !String

NCallGroup
emits a tag

Fields

nName :: !String

NCallSnd

Fields

nName :: !String

taskId :: !Int

NRetSnd

Fields

nName :: !String

taskId :: !Int

NRet

Fields

nName :: !String

NTagVal

Fields

nName :: !String

NValTag

Fields

nName :: !String

NCpHToDev

Fields

nName :: !String

NCpDevToH

Fields

nName :: !String

NCommit
2 outputs

Fields

nName :: !String

NStopSpec
2 outputs

Fields

nName :: !String

NArg

Fields

nName :: !String

NSuper

Fields

nName :: !String

superNum :: !Int

superOuts :: !Int

superImm :: !(Maybe Int)

superSpec :: !Bool

nodeName :: DNode -> String

Get the node's display/name label.

nOutputs :: DNode -> Int

Number of outputs (arity) for a given node.

outPort :: NodeId -> Port

Default primary output port "0".

out1Port :: NodeId -> Port

Secondary output port "1" (for multi-output nodes).

truePort :: NodeId -> Port

True branch port of a NSteer node.

falsePort :: NodeId -> Port

False branch port of a NSteer node.