Syntax

Description

Abstract Syntax Tree (AST) definitions for the simple functional language.

Documentation

type Ident = String

Identifier type, representing variable and function names.

data Program

A program is simply a list of declarations (functions).

Constructors

Program [Decl]

data Decl

A top-level declaration consists of a function name, its parameters, and its body expression.

Constructors

FunDecl Ident [Ident] Expr

data Expr

Expressions in the language.

Constructors

Var Ident
Variable reference

Lit Literal
Literal constant

Lambda [Ident] Expr
Lambda abstraction with parameters and body

If Expr Expr Expr
Conditional: if cond then then else else

Case Expr [(Pattern, Expr)]
Pattern matching: case expr of alternatives

Let [Decl] Expr
Local declarations and body expression

App Expr Expr
Function application

BinOp BinOperator Expr Expr
Binary operator application

UnOp UnOperator Expr
Unary operator application

List [Expr]
List literal

Tuple [Expr]
Tuple literal

data Pattern

Patterns used in case alternatives.

Constructors

PWildcard
Matches anything ("_" pattern)

PVar Ident
Variable pattern, binds the identifier

PLit Literal
Literal pattern

PList [Pattern]
List pattern

PTuple [Pattern]
Tuple pattern

data Literal

Literal values supported by the language.

Constructors

LInt Int
Integer literal

LFloat Double
Floating-point literal

LChar Char
Character literal

LString String
String literal

LBool Bool
Boolean literal

data BinOperator

Binary operators.

Constructors

Add
Addition (+)

Sub
Subtraction (-)

Mul
Multiplication (*)

Div
Division (/)

Mod
Modulus (mod)

Eq
Equality (==)

Neq
Inequality (/=)

Lt
Less than (<)

Le
Less than or equal (<=)

Gt
Greater than (>)

Ge
Greater than or equal (>=)

And
Logical and (&&)

Or
Logical or (||)

data UnOperator

Unary operators.

Constructors

Neg
Numeric negation (n -> -n)

Not
Logical negation (not)